requirements.txt
torch>=2.4.0
transformers>=4.45.0
datasets>=2.20.0
accelerate>=0.35.0
bitsandbytes>=0.44.0
tqdm>=4.66.0
numpy>=1.26.0
scipy>=1.13.0
pyyaml>=6.0.0
wandb>=0.17.0
# deepspeed and flash-attn require CUDA at build time.
# Install separately after GPU is confirmed working:
#   pip install packaging ninja && pip install deepspeed>=0.15.0
#   pip install flash-attn>=2.6.0 --no-build-isolation
